Add option to fetch missing artist images on manual Spotify response upload
